pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stardog-cloud-mcp"
version = "0.1.0"
description = "MCP Server for Stardog Cloud"
readme = "README.md"
requires-python = "~=3.12.0"
authors = [
    {name = "Stardog Union", email = "support@stardog.com"}
]
dependencies = [
    "fastmcp>=2.0.0",
    "pystardog==0.19.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "uvicorn>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "flake8"
]

[project.scripts]
stardog-cloud-mcp = "stardog_cloud_mcp.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/stardog_cloud_mcp"]

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true

[flake8]
max-line-length = 140